last upgrade for training level with code

+++ Final/Assets/Scripts/Target.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public class Target : MonoBehaviour
 {
-    public float health = 50f;
+    public int health = 50;
     //public GameObject NPC;
     public GameObject NPC_Ragdoll;
     
-    public void TakeDamage(float damage)
+    public void TakeDamage(int damage)
     {
         health -= damage;
         if(health == 0)
